support Either as constant in molecules
